Setting: utils, hooks폴더 추가, function폴더 삭제, pages폴더 내 하위폴더 삭제 (#34)
